# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Blog API",
    "description": "API for a simple blog application"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "description": "This endpoint registers a new user with a username, email, and password.",
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "User to register",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRegister"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User registered successfully",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in a user",
        "description": "This endpoint logs in a user by their email and password.",
        "parameters": [
          {
            "in": "body",
            "name": "credentials",
            "description": "User login credentials",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List all users",
        "description": "This endpoint returns a list of all users, excluding passwords and tokens.",
        "responses": {
          "200": {
            "description": "A list of users",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
    "/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "List all posts",
        "description": "This endpoint returns a list of all posts.",
        "responses": {
          "200": {
            "description": "A list of posts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Post"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Create a new post",
        "description": "This endpoint creates a new post with an author, title, and text.",
        "parameters": [
          {
            "in": "body",
            "name": "post",
            "description": "Post to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post created successfully",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/posts/{postId}": {
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete a post",
        "description": "This endpoint deletes a post by its ID.",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "Post deleted successfully"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/comments/{postId}": {
      "get": {
        "tags": ["Comments"],
        "summary": "List all comments for a post",
        "description": "This endpoint returns all comments associated with a post ID.",
        "parameters": [
          {
            "name": "postId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of comments",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Comment"
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Comments"],
        "summary": "Create a new comment on a post",
        "description": "This endpoint creates a new comment for a given post ID.",
        "parameters": [
          {
            "in": "body",
            "name": "comment",
            "description": "Comment to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment added successfully",
            "schema": {
              "$ref": "#/definitions/Comment"
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/comments/delete/{commentId}": {
      "delete": {
        "tags": ["Comments"],
        "summary": "Delete a comment",
        "description": "This endpoint deletes a comment by its ID.",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "204": {
            "description": "Comment deleted successfully"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "UserRegister": {
      "type": "object",
      "required": ["username", "email", "password"],
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "UserLogin": {
      "type": "object",
      "required": ["email", "password"],
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Post": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "author": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "date": {
          "type": "string",
          "format": "date-time"
        },
        "likes": {
          "type": "integer"
        },
        "dislikes": {
          "type": "integer"
        }
      }
    },
    "PostCreate": {
      "type": "object",
      "required": ["author", "title", "text"],
      "properties": {
        "author": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "text": {
          "type": "string"
        }
      }
    },
    "Comment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "postId": {
          "type": "integer",
          "format": "int64"
        },
        "text": {
          "type": "string"
        },
        "date": {
          "type": "string",
          "format": "date-time"
        },
        "likes": {
          "type": "integer"
        },
        "dislikes": {
          "type": "integer"
        }
      }
    },
    "CommentCreate": {
      "type": "object",
      "required": ["postId", "text"],
      "properties": {
        "postId": {
          "type": "integer",
          "format": "int64"
        },
        "text": {
          "type": "string"
        }
      }
    }
  }
}